begin code cleanup, fix naming to challenges
