+++ src/messages/phishes-answers.ts
@@ -1,0 +1,27 @@
+const phishesAnswers = `### 📧 Email 1: **Legitimate Email ✅**
+A real **password expiration notice** from the school IT department.  
+- **Key Features:** Uses a school-affiliated email (\`@school.edu\`), provides a **legitimate** password reset link, and offers **IT support contact options**.  
+- **No urgency or threats.**  
+
+---
+
+### 🚨 Email 2: **Phishing Email (Fake Security Alert) ❌**
+A **scam** pretending to be from IT Security, claiming your account has been "compromised."  
+- **Red Flags:** Uses **fear tactics**, a **fake deadline** (24 hours), and an **incorrect sender email** (\`@school-secure.com\`).  
+- **Goal:** Trick you into clicking a fake **login verification link** to steal credentials.  
+
+---
+
+### 🎓 Email 3: **Phishing Email (Fake Scholarship Offer) ❌**
+A **scam** pretending to offer a $1,000 scholarship.  
+- **Red Flags:** Too-good-to-be-true offer, **urgency tactic** ("before midnight tonight"), and a **non-school email domain**.  
+- **Goal:** Collect personal info via a fake "scholarship claim" link.  
+
+---
+
+### 🔍 **How to Spot Phishing Emails**
+- **Check the sender’s email:** Does it match the official domain?  
+- **Look for urgency tactics:** Phishers try to rush you into acting.  
+- **Verify links:** Hover over links before clicking to see the real URL.`
+
+export { phishesAnswers }
